feat: evm verify path and type
