+++ src/lib/components/evm-verify-section/FailedDetails.tsx
@@ -1,7 +1,7 @@
 import { Button, HStack, Text } from "@chakra-ui/react";
 
 import { useInternalNavigate, useMobile } from "lib/app-provider";
-import { EvmVerifyInfo } from "lib/services/types";
+import { EvmVerifyInfo } from "lib/types";
 import type { HexAddr20 } from "lib/types";
 import { formatUTC } from "lib/utils";
 import { EvmVerifyStatusModalWithTrigger } from "../modal/evm-verify-status";
