+++ src/lib/components/evm-verify-section/InProgressDetails.tsx
@@ -1,7 +1,7 @@
 import { Button, Text } from "@chakra-ui/react";
 
 import type { HexAddr20 } from "lib/types";
-import { EvmVerifyInfo } from "lib/services/types";
+import { EvmVerifyInfo } from "lib/types";
 import { EvmVerifyStatusModalWithTrigger } from "../modal/evm-verify-status";
 
 interface InProgressDetailsProps {
