+++ src/lib/pages/evm-contract-details/components/evm-contract-details-overview/index.tsx
@@ -14,7 +14,7 @@ import type { TxsTabIndex } from "../../types";
 import { EvmContractDetailsTxs } from "../EvmContractDetailsTxs";
 import { OverviewInfo } from "./OverviewInfo";
 import { OverviewVerifiedInfo } from "./OverviewVerifiedInfo";
-import { EvmVerifyInfo } from "lib/services/types";
+import { EvmVerifyInfo } from "lib/types";
 import { OverviewVerifiedCmds } from "./OverviewVerifiedCmds";
 
 interface EvmContractDetailsOverviewProps {
