+++ src/lib/pages/evm-contract-verify/components/evm-contract-verify-modal/EvmContractVerifyModalBody.tsx
@@ -4,7 +4,7 @@ import type { Control } from "react-hook-form";
 import { EvmContractVerifyFailedModal } from "./EvmContractVerifyFailedModal";
 import { EvmContractVerifyLoadingModal } from "./EvmContractVerifyLoadingModal";
 import { EvmContractVerifySuccessModal } from "./EvmContractVerifySuccessModal";
-import { EvmContractVerifyForm } from "lib/services/types";
+import { EvmContractVerifyForm } from "lib/types";
 
 interface EvmContractVerifyModalBodyProps {
   isError: boolean;
