+++ src/lib/pages/evm-contract-verify/components/vyper/EvmContractVerifyVyperContractCode.tsx
@@ -3,7 +3,7 @@ import { Heading, Stack } from "@chakra-ui/react";
 import { ConstructorArgs } from "../ConstructorArgs";
 import { EvmVersionToTarget } from "../EvmVersionToTarget";
 import { ControllerInput, ControllerTextarea } from "lib/components/forms";
-import { EvmContractVerifyForm, EvmVerifyConfig } from "lib/services/types";
+import { EvmContractVerifyForm, EvmVerifyConfig } from "lib/types";
 
 interface EvmContractVerifyVyperContractCodeProps {
   control: Control<EvmContractVerifyForm>;
@@ -56,11 +56,11 @@ export const EvmContractVerifyVyperContractCode = ({
       </Stack>
       <ConstructorArgs<EvmContractVerifyForm>
         control={control}
-        name="verifyForm.vyperContractCode"
+        name="verifyForm.vyperContractCode.constructorArgs"
       />
       <EvmVersionToTarget<EvmContractVerifyForm>
         control={control}
-        name="verifyForm.vyperContractCode"
+        name="verifyForm.vyperContractCode.evmVersion"
         evmVerifyConfig={evmVerifyConfig}
       />
     </Stack>
